also remove references to CGameContext from CServer
