+++ src/engine/shared/snapshot.h
@@ -45,6 +45,7 @@ public:
 
 	int Crc();
 	void DebugDump();
+	static void RemoveExtraInfo(unsigned char *pData);
 };
