+++ src/game/gamecore.h
@@ -296,9 +296,4 @@ inline CInputCount CountInput(int Prev, int Cur)
 	return c;
 }
 
-bool UseExtraInfo(const CNetObj_Projectile *pProj);
-void ExtractInfo(const CNetObj_Projectile *pProj, vec2 *StartPos, vec2 *StartVel, bool IsDDNet);
-void ExtractExtraInfo(const CNetObj_Projectile *pProj, int *Owner, bool *Explosive, int *Bouncing, bool *Freeze);
-void SnapshotRemoveExtraInfo(unsigned char *pData);
-
 #endif
