+++ src/game/server/gamecontext.h
@@ -226,6 +226,9 @@ public:
 	// Describes the time when the first player joined the server.
 	int64 m_NonEmptySince;
 	int64 m_LastMapVote;
+	int GetClientVersion(int ClientID);
+	void SetClientVersion(int ClientID, int Version);
+	bool PlayerExists(int ClientID) { return m_apPlayers[ClientID]; };
 
 private:
